third_party/blink/web_tests/external/wpt/css/css-box/margin-trim/computed-margin-values/block-container-block-start-child-with-border-expected.txt

This is a testharness.js-based test.
[FAIL] item 1
  assert_equals: \n<item data-expected-margin-top="0" class="collapsed">\n            <item data-expected-margin-top="0" class="collapsed"></item>\n        </item>\nmargin-top expected "0" but got "50"
[FAIL] item 2
  assert_equals: \n<item data-expected-margin-top="0" class="collapsed"></item>\nmargin-top expected "0" but got "50"
[FAIL] item 3
  assert_equals: \n<item class="with-border-and-margin" data-expected-margin-top="0">\n            <item class="with-border-and-margin" data-expected-margin-top="30"></item>\n        </item>\nmargin-top expected "0" but got "30"
[FAIL] item 4
  assert_equals: \n<item class="with-border-and-margin" data-expected-margin-top="30"></item>\nmargin-top expected "0" but got "30"
Harness: the test ran to completion.
